remove unused code and files
delete reducers/sorting.js, actions/sorting.js had searching actions also so rename sorting.js to search.js and delete sorting actions from it, delete getFilteredCollections and getSortedCollections from selectors/collections.js, delete selectors/projects.js, delete TableRow.jsx
