+++ client/modules/IDE/actions/search.js
@@ -1,0 +1,13 @@
+import * as ActionTypes from '../../../constants';
+
+export function setSearchTerm(scope, searchTerm) {
+  return {
+    type: ActionTypes.SET_SEARCH_TERM,
+    query: searchTerm,
+    scope
+  };
+}
+
+export function resetSearchTerm(scope) {
+  return setSearchTerm(scope, '');
+}
